Add a deadzone option to joyosc.pd.
A minimum deadzone of 1 around the center position of the sticks is now
always enforced, which can be enlarged to 129 by turning on the deadzone
toggle in the main patch. This is most useful when working with TouchOSC
which often seems to report nonzero center positions for the sticks.

I'm not sure whether that's a bug in TouchOSC itself, or the SDL2
version it uses, and the precise amount of deviation seems to vary. But
the increased deadzone of 129 should hopefully work with most modern
gamepads and alleviate the need for a calibration (that option is still
available as well, however).

Note that none of this is needed when using joyosc, which properly
reports center positions as (0,0), at least with all the gamepads I
tried it with.